refactor: Move all the source code into src folder
